+++ src/modules/pagination/pagination.js
@@ -3,10 +3,11 @@ import Pagination from 'tui-pagination';
 import 'tui-pagination/dist/tui-pagination.min.css';
 import renderMoviesList from '../html-render';
 import { BASE_URL, API_KEY } from '../api-service';
+import fetchApi, { apiRefs } from '../api-service';
 
 // let paginationActive = false;
 
-const container = refs.paginationRef;
+// const container = refs.paginationRef;
 const searchQueryParameter = 'ukraine';
 
 const searchQueryOptions = {
@@ -15,32 +16,33 @@ const searchQueryOptions = {
   totalPages: 7,
 };
 
-const pagination = new Pagination(container, {
-  totalItems: searchQueryOptions.totalPages * 20, // Assuming 20 items per page
-  itemsPerPage: 20, // Number of items to display per page
-  visiblePages: 5, // Number of visible pages in the pagination
-  currentPage: searchQueryOptions.page, // Current page
-  centerAlign: true,
-});
+// const pagination = new Pagination(container, {
+//   totalItems: searchQueryOptions.totalPages * 20, // Assuming 20 items per page
+//   itemsPerPage: 20, // Number of items to display per page
+//   visiblePages: 5, // Number of visible pages in the pagination
+//   currentPage: searchQueryOptions.page, // Current page
+//   centerAlign: true,
+// });
 
-pagination.on('afterMove', eventData => {
-  searchQueryOptions.page = eventData.page;
+// pagination.on('afterMove', eventData => {
+//   searchQueryOptions.page = eventData.page;
 
-  performMovieSearch();
-});
+//   performMovieSearch();
+// });
 
-function performMovieSearch() {
-  const searchURL = `${searchQueryOptions.searchURL}&page=${searchQueryOptions.page}`;
-  fetch(searchURL)
-    .then(response => response.json())
-    .then(data => {
-      // Update the UI with the movie search results
-      renderMoviesList(data.results);
-    })
-    .catch(error => {
-      console.error('Error performing movie search:', error);
-    });
-}
+// function performMovieSearch() {
+//   const searchURL = `${searchQueryOptions.searchURL}&page=${searchQueryOptions.page}`;
+//   fetch(searchURL)
+//     .then(response => response.json())
+//     .then(data => {
+//       // Update the UI with the movie search results
+//       renderMoviesList(data.results);
+//     })
+//     .catch(error => {
+//       console.error('Error performing movie search:', error);
+//     });
+// }
+
 // if (paginationActive) {
 //   const pagination = new Pagination(container, options);
 // } else {
