+++ src/modules/search/fetch.js
@@ -1,17 +1,26 @@
 import { refs } from '../refs';
 import { Notify } from 'notiflix';
-import fetchApi from '../api-service';
+import fetchApi, { apiRefs } from '../api-service';
 import { apiRefs } from '../api-service';
 import renderMoviesList from '../html-render';
-
+import Pagination from 'tui-pagination';
+import 'tui-pagination/dist/tui-pagination.min.css';
+import { BASE_URL, API_KEY } from '../api-service';
 
 // ref.alertMessage = document.querySelector('.header-form--search'),
 refs.formRef.addEventListener('submit', fetchFilmOfSearch);
 
+let pagination;
+
 export async function fetchFilmOfSearch(e) {
   e.preventDefault();
+
   const searchQuery = e.currentTarget.elements.film.value.trim();
-  e.currentTarget.reset(); 
+  e.currentTarget.reset();
+
+  if (pagination) {
+    pagination.reset();
+  }
 
   const data = await fetchApi({
     param: apiRefs.SEARCH,
@@ -22,27 +31,57 @@ export async function fetchFilmOfSearch(e) {
   console.log(dataEl);
   console.log(data);
 
+  if (searchQuery === '') {
+    refs.alertMessage.style.display = 'none';
+    Notify.warning('Searching starts after providing data to search!');
+    return;
+  } else if (dataEl === 0) {
+    // refs.alertMessage.classList.add('visually-hidden');
+    refs.alertMessage.style.display = 'flex';
+    Notify.warning(
+      'Search result not successful. Enter the correct movie name!'
+    );
+    return;
+  } else {
+    refs.alertMessage.style.display = 'none';
+    Notify.success(` We found ${data.total_results} films.`);
+  }
+  renderMoviesList(data.results);
 
-if (searchQuery === '') {
-  refs.alertMessage.style.display = 'none';
-  Notify.warning('Searching starts after providing data to search!');
-  return;
-} else if (dataEl === 0) {
-  // refs.alertMessage.classList.add('visually-hidden');
-  refs.alertMessage.style.display = 'flex';
-  Notify.warning('Search result not successful. Enter the correct movie name!');
-  return;
+  const paginationParams = {
+    totalItems: data.total_results,
+    itemsPerPage: 20,
+    visiblePages: 5,
+    currentPage: 1,
+    centerAlign: true,
+  };
 
-} else {
-  refs.alertMessage.style.display = 'none';
-  Notify.success(` We found ${data.total_results} films.`);
+  pagination = new Pagination(refs.paginationRef, paginationParams);
+
+  pagination.on('afterMove', eventData => {
+    paginationParams.currentPage = eventData.page;
+
+    performMovieSearch();
+  });
+
+  function performMovieSearch() {
+    fetchApi({
+        param: apiRefs.SEARCH,
+        page: paginationParams.currentPage,
+        query: searchQuery,
+      })
+    // const searchURL = `${BASE_URL}/search/movie?api_key=${API_KEY}&query=${searchQuery}&page=${paginationParams.currentPage}`;
+    // fetch(searchURL)
+      // .then(response => response.json())
+      .then(data => {
+        // Update the UI with the movie search results
+        renderMoviesList(data.results);
+      })
+      .catch(error => {
+        console.error('Error performing movie search:', error);
+      });
+  }
 }
-renderMoviesList(data.results);
-}
-
-
-
-
 
 // refs.formRef.addEventListener('submit', fetchFilmOfSearch);
